Highlight current bookmark status
TODO: swap out dropdown with htmx to reflect change without page refresh

+++ src/main/java/com/swdev/springbootproject/service/TMDBService.java
@@ -95,7 +95,7 @@ public class TMDBService {
     return discoverResults.getResults();
   }
 
-  public MovieDetails getMovieDetails(int movieId) {
+  public MovieDetails getMovieDetails(Long movieId) {
     return restClient
         .get()
         .uri(
